Clean up and test
