+++ Cargo.toml
@@ -79,6 +79,8 @@ rand = "0.8.5"
 reqwest = { version = "0.12.0", features = ["blocking"] }
 seq-macro = "0.3.5"
 serde = "1.0.197"
+serde_json = "1.0.116"
+serde_test = "1.0.176"
 simplelog = { version = "0.12.2", features = ["paris"] }
 thiserror = "1.0.58"
 tokio = "1.37.0"
