+++ vortex-array/src/array/sparse/compute/mod.rs
@@ -138,7 +138,8 @@ fn take_search_sorted(
 #[cfg(test)]
 mod test {
     use itertools::Itertools;
-    use vortex_dtype::{DType, Nullability, PType};
+    use vortex_dtype::Nullability;
+    use vortex_dtype::{DType, PType};
     use vortex_scalar::Scalar;
 
     use crate::array::primitive::PrimitiveArray;
