+++ vortex-dict/src/compute.rs
@@ -52,7 +52,8 @@ mod test {
     use vortex::array::primitive::PrimitiveArray;
     use vortex::array::varbin::VarBinArray;
     use vortex::{IntoArray, ToArray};
-    use vortex_dtype::{DType, Nullability};
+    use vortex_dtype::DType;
+    use vortex_dtype::Nullability;
 
     use crate::{dict_encode_typed_primitive, dict_encode_varbin, DictArray};
