+++ vortex-dtype/src/nullability.rs
@@ -1,0 +1,36 @@
+use std::fmt::{Display, Formatter};
+
+#[derive(Debug, Clone, Copy, Default, PartialEq, Eq, Hash, Ord, PartialOrd)]
+pub enum Nullability {
+    #[default]
+    NonNullable,
+    Nullable,
+}
+
+impl From<bool> for Nullability {
+    fn from(value: bool) -> Self {
+        if value {
+            Nullability::Nullable
+        } else {
+            Nullability::NonNullable
+        }
+    }
+}
+
+impl From<Nullability> for bool {
+    fn from(value: Nullability) -> Self {
+        match value {
+            Nullability::NonNullable => false,
+            Nullability::Nullable => true,
+        }
+    }
+}
+
+impl Display for Nullability {
+    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
+        match self {
+            Nullability::NonNullable => write!(f, ""),
+            Nullability::Nullable => write!(f, "?"),
+        }
+    }
+}
