--- vortex-dtype/src/serde.rs
@@ -1,1 +1,0 @@
-#![cfg(feature = "serde")]
