+++ vortex-ipc/benches/ipc_array_reader_take.rs
@@ -5,7 +5,8 @@ use fallible_iterator::FallibleIterator;
 use itertools::Itertools;
 use vortex::array::primitive::PrimitiveArray;
 use vortex::{Context, IntoArray};
-use vortex_dtype::{DType, Nullability, PType};
+use vortex_dtype::Nullability;
+use vortex_dtype::{DType, PType};
 use vortex_ipc::iter::FallibleLendingIterator;
 use vortex_ipc::reader::StreamReader;
 use vortex_ipc::writer::StreamWriter;
