+++ vortex-ree/src/ree.rs
@@ -145,7 +145,8 @@ mod test {
     use vortex::compute::slice::slice;
     use vortex::validity::Validity;
     use vortex::{ArrayDType, ArrayTrait, IntoArray};
-    use vortex_dtype::{DType, Nullability, PType};
+    use vortex_dtype::Nullability;
+    use vortex_dtype::{DType, PType};
 
     use crate::REEArray;
